# src/doc/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Edugram API",
    "description": "Edugram Backend API",
    "version": "1.0.0"
  },
  "host": "katowice.edugram.io",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/api/author/": {
      "get": {
        "tags": [
          "Author"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "Get Author details",
            "schema": {
              "type": "object",
              "properties": {
                "firstName": {
                  "type": "string",
                  "example": "John"
                },
                "lastName": {
                  "type": "string",
                  "example": "Doe"
                },
                "avatar": {
                  "type": "string",
                  "example": "https://cdn.edugram.io/user-storage/12345/avatars/example.png"
                },
                "averageRating": {
                  "type": "number",
                  "example": 4.85
                },
                "totalStudents": {
                  "type": "number",
                  "example": 100
                },
                "totalCourses": {
                  "type": "number",
                  "example": 8
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "status": {
                  "type": "number",
                  "example": 404
                },
                "message": {
                  "type": "string",
                  "example": "User not found"
                },
                "stack": {
                  "type": "object",
                  "properties": {}
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Not Found"
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "status": {
                  "type": "number",
                  "example": 500
                },
                "message": {
                  "type": "string",
                  "example": "Something went wrong"
                },
                "stack": {
                  "type": "object",
                  "properties": {}
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/author/courses": {
      "get": {
        "tags": [
          "Author"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "Get Author courses",
            "schema": {
              "type": "array",
              "example": [
                "courses"
              ],
              "items": {
                "type": "string"
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "status": {
                  "type": "number",
                  "example": 500
                },
                "message": {
                  "type": "string",
                  "example": "Something went wrong"
                },
                "stack": {
                  "type": "object",
                  "properties": {}
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/payments/checkout": {
      "post": {
        "tags": [
          "Payments"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "items": {
                  "example": "any"
                },
                "discounts": {
                  "example": "any"
                },
                "currency": {
                  "example": "any"
                },
                "itemType": {
                  "example": "any"
                },
                "orderId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "client_secret": {
                  "type": "string",
                  "example": ""
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "201": {
            "schema": {
              "$ref": "#/definitions/payment/properties/201"
            },
            "description": "Created"
          },
          "500": {
            "schema": {
              "$ref": "#/definitions/payment/properties/500"
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/payments/discount": {
      "post": {
        "tags": [
          "Payments"
        ],
        "description": "",
        "parameters": [
          {
            "name": "isGiftOrder",
            "in": "query",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "example": "any"
                },
                "items": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/coupon/properties/CouponDetails"
            }
          },
          "500": {
            "schema": {
              "$ref": "#/definitions/coupon/properties/500"
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/categories/": {
      "get": {
        "tags": [
          "Categories"
        ],
        "description": "",
        "parameters": [
          {
            "name": "accept-language",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Categories"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/categories/{category}": {
      "get": {
        "tags": [
          "Categories"
        ],
        "description": "",
        "parameters": [
          {
            "name": "category",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "accept-language",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/categories/{categoryId}": {
      "patch": {
        "tags": [
          "Categories"
        ],
        "description": "",
        "parameters": [
          {
            "name": "categoryId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Categories"
        ],
        "description": "",
        "parameters": [
          {
            "name": "categoryId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
...
}